email-service/requirements.txt
supabase==2.3.0
resend==0.7.0
python-dotenv==1.0.0
schedule==1.2.1
jinja2==3.1.2
